Moving to string resource usage
