+++ src/modules/peek/Peek.FilePreviewer/Previewers/Helpers/IconHelper.cs
@@ -2,7 +2,7 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Peek.FilePreviewer.Previewers
+namespace Peek.FilePreviewer.Previewers.Helpers
 {
     using System;
     using System.Drawing;
