+++ src/modules/peek/Peek.FilePreviewer/Previewers/Helpers/ReadableStringHelper.cs
@@ -2,10 +2,11 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Peek.FilePreviewer.Previewers
+namespace Peek.FilePreviewer.Previewers.Helpers
 {
     using System;
     using System.Collections.Generic;
+    using Windows.ApplicationModel.Resources;
 
     public static class ReadableStringHelper
     {
@@ -13,16 +14,16 @@ namespace Peek.FilePreviewer.Previewers
 
         public static string BytesToReadableString(ulong bytes)
         {
-            // TODO: get string from resources
+            var resourceLoader = ResourceLoader.GetForViewIndependentUse();
             List<string> format = new List<string>
             {
-                "B",
-                "KB",
-                "MB",
-                "GB",
-                "TB",
-                "PB",
-                "EB",
+                resourceLoader.GetString("ReadableString_ByteAbbreviationFormat"),     // "B"
+                resourceLoader.GetString("ReadableString_KiloByteAbbreviationFormat"), // "KB"
+                resourceLoader.GetString("ReadableString_MegaByteAbbreviationFormat"), // "MB"
+                resourceLoader.GetString("ReadableString_GigaByteAbbreviationFormat"), // "GB"
+                resourceLoader.GetString("ReadableString_TeraByteAbbreviationFormat"), // "TB"
+                resourceLoader.GetString("ReadableString_PetaByteAbbreviationFormat"), // "PB"
+                resourceLoader.GetString("ReadableString_ExaByteAbbreviationFormat"),  // "EB"
             };
 
             int index = 0;
@@ -34,7 +35,15 @@ namespace Peek.FilePreviewer.Previewers
                 number = Math.Round((bytes / Math.Pow(1024, index)) * DecimalPercision) / DecimalPercision;
             }
 
-            return string.Concat(number, format[index]);
+            return string.Format(format[index], number);
+        }
+
+        public static string FormatResourceString(string resourceId, object? args)
+        {
+            var formatString = ResourceLoader.GetForViewIndependentUse().GetString(resourceId);
+            var formattedString = string.Format(formatString, args);
+
+            return formattedString;
         }
     }
 }
