+++ src/modules/peek/Peek.FilePreviewer/Previewers/UnsupportedFilePreviewer/UnsupportedFilePreviewer.cs
@@ -13,6 +13,7 @@ namespace Peek.FilePreviewer.Previewers
     using Microsoft.UI.Dispatching;
     using Microsoft.UI.Xaml.Media.Imaging;
     using Peek.Common;
+    using Peek.FilePreviewer.Previewers.Helpers;
     using Windows.Foundation;
     using File = Peek.Common.Models.File;
